完善系统初始化与ID序列化处理

+++ eva-core/eva-core-data-mybatis/src/main/java/org/pkaq/core/mybatis/mvc/service/StdTreeService.java
@@ -115,7 +115,7 @@ public abstract class StdTreeService<M extends StdTreeMapper<T>, T extends StdTr
             orderQuery.eq(T::getPid, pid);
             orderQuery.eq(T::getDeleted, DelEnumm.UN_DELETED);
 
-            entity.setSort(this.mapper.selectCount(orderQuery));
+            entity.setSort((double) this.mapper.selectCount(orderQuery));
         } else {
             oldOrgin = this.mapper.selectById(orgId);
         }
